@@ -47,7 +47,7 @@ class TestGeoipsDescribePackage(BaseCliTest):
         """
         # An error occurred using args. Assert that args is not valid and check the
         # output of the error.
-        err_str = "usage: To use, type `geoips describe package <package_name>`"
+        err_str = "usage: To use, type `geoips describe package <package-name>`"
         assert err_str in error
 
     def check_output(self, args, output):
@@ -62,7 +62,7 @@ class TestGeoipsDescribePackage(BaseCliTest):
         """
         # The args provided are valid, so test that the output is actually correct
         if "-h" in args:
-            usg_str = "usage: To use, type `geoips describe package <package_name>`"
+            usg_str = "usage: To use, type `geoips describe package <package-name>`"
             assert usg_str in output
         else:
             # Checking that output from geoips describe package command is valid
